Stopwatch, Distance, and Elevation implemented
Thanks to Dean for writing much of this code, he sent it to me via discord!

The DrillStopwatch contains a StreamController with a stream that increments every second, and now DuringDrill contains some clever code to display the counter as minutes:seconds. Thanks Dean!

I moved the LocationTracker from BasicDrillPresenter to DuringDrill so that we could extend it with state variables distanceTravelled and currentElevation.

I pulled code from Dean which calculates the distance travelled into LocationTracker. Thanks Dean!

Every tenth ParticipantLocation (about one per second) updates the distanceTravelled and currentElevation, and the widgets depend on that state so they update at the same time.

Huzzah!